+++ src/main/java/roomescape/controller/TimeController.java
@@ -1,0 +1,4 @@
+package roomescape.controller;
+
+public class TimeController {
+}
