+++ src/main/java/roomescape/entity/Time.java
@@ -1,0 +1,4 @@
+package roomescape.entity;
+
+public class TIme {
+}
